refactor; moved commands
