+++ cognite_toolkit/_cdf_tk/apps/__init__.py
@@ -1,9 +1,9 @@
 from ._auth_app import AuthApp
 from ._core_app import CoreApp
+from ._dump_app import DumpApp
 from ._landing_app import LandingApp
 from ._modules_app import ModulesApp
 from ._repo_app import RepoApp
 from ._run import RunApp
-from .dump_app import DumpApp
 
-__all__ = ["CoreApp", "AuthApp", "LandingApp", "ModulesApp", "RunApp", "RepoApp", "DumpApp"]
+__all__ = ["CoreApp", "AuthApp", "LandingApp", "DumpApp", "ModulesApp", "RunApp", "RepoApp"]
