+++ cognite_toolkit/_cdf_tk/commands/dump_assets.py
