--- cognite_toolkit/_cdf_tk/prototypes/commands/__init__.py
@@ -1,4 +1,0 @@
-from .dump_assets import DumpAssetsCommand
-from .dump_timeseries import DumpTimeSeriesCommand
-
-__all__ = ["DumpAssetsCommand", "DumpTimeSeriesCommand"]
